// CasualGameMobile/types.ts
export type PieceType = 'BREAD' | 'MEAT' | 'CHEESE' | 'LETTUCE' | 'TOMATO' | 'BACON' | 'KETCHUP' | 'PICKLE' | 'ONION';
export type Screen = 'MENU' | 'INTRO' | 'GAME' | 'RESULT';
export type GameMode = 'CAMPAIGN' | 'ARCADE';

export interface Piece {
  id: string;
  type: PieceType;
  isNew?: boolean;
  isRemoving?: boolean;
}

export interface Cell {
  row: number;
  col: number;
  piece: Piece | null;
}

export interface Level {
  id: number;
  name: string;
  targetMoney: number;
  ingredients: PieceType[];
  newIngredient: PieceType;
  description: string;
}